+++ whatsNew.md
@@ -2,5 +2,7 @@ Content :
 * Libs : all DLL libraries
 * Windows Forms Example
 * Linux, Mac OS and Windows console example
+* Node.js example
+* Python example
 
 By downloading, you hereby agree to the [terms and conditions](https://underautomation.com/eula)